pyproject.toml
[tool.poetry]
name = "education"
version = "0.1.0"
description = "Education website"
authors = ["sudodebug <sudodebug@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
pillow = "^10.0.0"
psycopg2-binary = "^2.9.6"
django-braces = "^1.15.0"
django-embed-video = "^1.4.9"
pymemcache = "^4.0.0"
django-debug-toolbar = "^4.1.0"
redis = "^4.6.0"
django-redisboard = "^8.4.0"
attrs = "^23.1.0"
djangorestframework = "^3.14.0"
requests = "^2.31.0"
channels = "3.0.5"
channels-redis = "3.4.1"
flake8 = "^6.1.0"
black = "^23.9.1"
uwsgi = "^2.0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
